Various initial cleanup
