+++ packages/engine/src/gltf/GLTFLoaderUtils.ts
@@ -2,10 +2,8 @@ import {
   Box3,
   BufferAttribute,
   ColorManagement,
-  FrontSide,
   Interpolant,
   LinearSRGBColorSpace,
-  MeshStandardMaterial,
   Quaternion,
   Sphere,
   TriangleFanDrawMode,
@@ -409,25 +407,6 @@ export function addPrimitiveAttributes(geometry, primitiveDef, parser) {
   return Promise.all(pending).then(function () {
     return primitiveDef.targets !== undefined ? addMorphTargets(geometry, primitiveDef.targets, parser) : geometry
   })
-}
-
-/**
- * Specification: https://github.com/KhronosGroup/glTF/blob/master/specification/2.0/README.md#default-material
- */
-export function createDefaultMaterial(cache) {
-  if (cache['DefaultMaterial'] === undefined) {
-    cache['DefaultMaterial'] = new MeshStandardMaterial({
-      color: 0xffffff,
-      emissive: 0x000000,
-      metalness: 1,
-      roughness: 1,
-      transparent: false,
-      depthTest: true,
-      side: FrontSide
-    })
-  }
-
-  return cache['DefaultMaterial']
 }
 
 export function addUnknownExtensionsToUserData(knownExtensions, object, objectDef) {
